+++ libft/libft.h
@@ -64,7 +64,7 @@ char	*ft_strnstr(const char *big, const char *little, size_t len);
 
 int		ft_atoi(const char *nptr);
 
-long	ft_atoll(const char *nptr);
+long long	ft_atoll(const char *nptr);
 
 void	*ft_calloc(size_t nmemb, size_t size);
